spam-prediction-api/requirements.txt
google-cloud-storage 
scikit-learn 
xgboost
 scipy 
functions-framework